macOS: Simulated input is independent of the physical keyboard state
Previously the state of the physical keyboard effected the result of the simulated input. If the Command key was held on the physical keyboard, the text() function no longer worked. By default the two are now independent, but this behavior can be changed with the Settings struct (`independent_of_keyboard_state`).
Fixes https://github.com/enigo-rs/enigo/issues/297