macOS: Remove all sleeps and add one when the `Enigo` struct is dropped
It looks like all pending events get ignored when the program terminates. It seems like the sleeps were only needed so that the OS had enough time to handle all events before the struct gets dropped. Sleeps during the execution are much more annoying then at the end of the program, so I just added a sleep of two seconds when Enigo gets dropped. That is more than enough for most cases.