macos: Only sleep when needed
All events need to be handled by the OS before the Enigo struct gets dropped. Previously a sleep of two seconds was used to ensure that. This was too long when only a few events were sent or if a long time has passed since sending the events. The sleep time is now calculated. We assume the OS needs 20 ms to handle each event. For each event we send, we calculate how long we need to wait to ensure all previous events were handled and add 20 ms of waiting time