linux: wayland: Sync the state of the compositor when establishing a connection
The previous implementation was very sequential and blocked multiple times until it received an event from the compositor. This lead to infinite hangs on some compositors (see https://github.com/enigo-rs/enigo/issues/396). These cases were replaced with calls to synchronize the compositor with the client. Now the library will only freeze, if the compositor never sends a "Done" event to signal that the states are synchronized. These cases would be bugs in the compositors code though.