linux: wayland: Use keymap received from the compositor
We previously used a hardcoded keymap for the virtual_keyboard protocol. With this commit, the current keymap that is received from the compositor is used. That keymap is more extensive and it has the benefit that simulating raw keycodes has the same effect as if they were sent from a physical device. If a user wants to simulate a specific keysym, they should not use the raw() function anyways.
Another benefit is that the modifiers now work as expected and not only eight modifers work. The keymap dictates which keys are modifiers so using the keymap from the compositor makes all modifiers work just as they were mapped.
The important downside of this commit is the large amount of code added. It is needed, because the xkbcommon library does not provide functions to mutate the keymap. We need to do that though to map keysyms to unused keycodes. Otherwise only the keysyms which are currently mapped can be simulated. The parsing is done to find unused keycodes and unused identifiers.
There is probably a small performance hit because of this, but I did not try to measure it. It was not noticeable in my tests.