linux: x11: Change the default feature to x11rb
It is more reliable, makes it possible to enter raw keycodes, does not need `xdotools` as a runtime dependency and uses more native Rust. If you experience issue with the new default, please open an issue. You can disable default features and use the feature `xdo` to use the old method