win: Fix extended keys
The scan code of extended keys starts with 0xE0 or 0xE1. If that is the case, the proper KEYBD_EVENT_FLAGS need to be set. The code now properly detects extended keys
in the raw() function. In addition some keys get mapped to the same virtual key. That is why the virtual key cannot be used to determine if it is an extended key.
Instead the enigo::Key enum is used. The enigo::EXT constant was incorrect and removed.
The `NumpadEnter` key was added on Linux and Windows.
Fixes https://github.com/enigo-rs/enigo/issues/476