tests: Browser sends message with detected input to test
The integration tests were complex and brittle. Instead of the browser and the test
suite sending messages to each other, it was changed so only the browser sends
messages. The messages contain the detected events as little changed as reasonable.
All logic of comparing expected events with the received events happens on the Rust side.