all: Use the "key" field of Firefox events as a fallback
It is done because on Windows (maybe other platforms too), "code" is empty for the "Help" key. The "key" field does contain the correct value though. We cannot always use the "key" field, because it would make it impossible to differentiate left and right keys (e.g LControl from RControl).