macos: fix crash by executing TIS keyboard layout retrieval on the main thread
All TIS functions must be called from the main thread. When using Tauri and enigo together, this was often not the case, so the application crashed. To fix this issue, we use the dispatch2::run_on_main function to execute the TIS function on main.
This only works if the there is for example a CFRunLoop. Otherwise it will hang. When the enigo struct is created, we try to detect if the requirement is fulfilled. If it is, we use the run_on_main function. If it is not, we don't use it